[UPDATE] 오류 수정 및 코드 컴포넌트화
