+++ src/components/message/ChatDate.tsx
@@ -1,0 +1,36 @@
+import { format, parseISO } from 'date-fns';
+import Image from 'next/image';
+
+export function GetCurrentTime(time: string | undefined) {
+  if (time) {
+    const formattedTime = format(parseISO(time), 'p');
+    return formattedTime;
+  }
+}
+
+export function DisplayDateTime(date: string) {
+  if (date) {
+    const formattedDate = format(parseISO(date), 'yyyy년 MM월 dd일');
+    return (
+      <div className="flex mr-auto ml-auto justify-center mt-[1rem]">
+        <div className="flex justify-center items-center border-1 border-solid border-gray-DDD  py-[0.5rem] rounded-[3.125rem] text-[0.875rem] px-[1rem] text-gray-999">
+          <Image
+            src="/assets/figmaImg/Calendar.png"
+            className="w-[1rem] h-[1rem] mr-[0.25rem]"
+            width={50}
+            height={50}
+            alt="캘린더"
+          />
+          {formattedDate}
+        </div>
+      </div>
+    );
+  }
+}
+
+export function ConvertedDate(date: string | undefined, idx: number) {
+  if (date) {
+    const formattedDate = format(parseISO(date), 'P');
+    return formattedDate;
+  }
+}
